Added test scripts for Tickets 508 & 497
